requirements.txt
mitmproxy==10.3.0
pydantic==2.7.1
requests==2.31.0
PyYAML==6.0.1
watchdog==4.0.0
uvicorn==0.29.0
typer==0.12.3
python-dotenv==1.0.1